modif dao comAer

+++ src/main/java/tpVol/dao/DaoCompagnieAerienne.java
@@ -2,6 +2,6 @@ package tpVol.dao;
 
 import tpVol.model.CompagnieAerienne;
 
-public interface DaoCompagnieAerienne extends DaoGeneric<CompagnieAerienne, Integer>{
+public interface DaoCompagnieAerienne extends DaoGeneric<CompagnieAerienne, Long>{
 	
 }
